requirements.txt
whitenoise==6.2.0
Pillow==9.1.0

django-phonenumber-field==6.3.0
phonenumbers==8.12.54
Babel==2.13.0
